+++ include/swift/SIL/InstructionUtils.h
@@ -2,7 +2,7 @@
 //
 // This source file is part of the Swift.org open source project
 //
-// Copyright (c) 2014 - 2017 Apple Inc. and the Swift project authors
+// Copyright (c) 2014 - 2018 Apple Inc. and the Swift project authors
 // Licensed under Apache License v2.0 with Runtime Library Exception
 //
 // See https://swift.org/LICENSE.txt for license information
@@ -115,23 +115,6 @@ bool onlyAffectsRefCount(SILInstruction *user);
 /// recursively.
 SILValue stripConvertFunctions(SILValue V);
 
-/// Given an address accessed by an instruction that reads or modifies
-/// memory, return the base address of the formal access. If the given address
-/// is produced by an initialization sequence, which cannot correspond to a
-/// formal access, then return an invalid SILValue.
-///
-/// This must return a valid SILValue for the address operand of begin_access.
-SILValue findAccessedAddressBase(SILValue sourceAddr);
-
-/// Return true if the given address producer may be the source of a formal
-/// access (a read or write of a potentially aliased, user visible variable).
-///
-/// If this returns false, then the address can be safely accessed without
-/// a begin_access marker. To determine whether to emit begin_access:
-///   base = findAccessedAddressBase(address)
-///   needsAccessMarker = base && baseAddressNeedsFormalAccess(base)
-bool isPossibleFormalAccessBase(SILValue baseAddress);
-
 /// Check that this is a partial apply of a reabstraction thunk and return the
 /// argument of the partial apply if it is.
 SILValue isPartialApplyOfReabstractionThunk(PartialApplyInst *PAI);
@@ -147,13 +130,6 @@ struct LLVM_LIBRARY_VISIBILITY FindClosureResult {
 /// IsReabstractionThunk flag set to true if the closure is indirectly captured
 /// by a reabstraction thunk.
 FindClosureResult findClosureForAppliedArg(SILValue V);
-
-/// Visit each address accessed by the given memory operation.
-///
-/// This only visits instructions that modify memory in some user-visible way,
-/// which could be considered part of a formal access.
-void visitAccessedAddress(SILInstruction *I,
-                          std::function<void(Operand *)> visitor);
 
 /// A utility class for evaluating whether a newly parsed or deserialized
 /// function has qualified or unqualified ownership.
