+++ lib/SIL/InstructionUtils.cpp
@@ -313,171 +313,6 @@ SILValue swift::stripConvertFunctions(SILValue V) {
   return V;
 }
 
-// Return true if the given address is a 'let' lvalue.
-static bool isLetAccess(SILValue address) {
-  switch (address->getKind()) {
-  default:
-    return false;
-
-  case ValueKind::AllocStackInst: {
-    VarDecl *decl = cast<AllocStackInst>(address)->getDecl();
-    return decl && decl->isLet();
-  }
-  case ValueKind::AllocBoxInst: {
-    VarDecl *decl = cast<AllocBoxInst>(address)->getDecl();
-    return decl && decl->isLet();
-  }
-  case ValueKind::RefElementAddrInst: {
-    VarDecl *decl = cast<RefElementAddrInst>(address)->getField();
-    return decl && decl->isLet();
-  }
-  case ValueKind::GlobalAddrInst: {
-    SILGlobalVariable *global =
-        cast<GlobalAddrInst>(address)->getReferencedGlobal();
-    return global && global->isLet();
-  }
-  };
-}
-
-// An address base is a block argument. Verify that it is actually a box
-// projected from a switch_enum.
-static void checkSwitchEnumBlockArg(SILPHIArgument *arg) {
-  assert(!arg->getType().isAddress());
-  SILBasicBlock *Pred = arg->getParent()->getSinglePredecessorBlock();
-  if (!Pred || !isa<SwitchEnumInst>(Pred->getTerminator())) {
-    arg->dump();
-    llvm_unreachable("unexpected box source.");
-  }
-}
-
-SILValue swift::findAccessedAddressBase(SILValue sourceAddr) {
-  SILValue address = sourceAddr;
-  while (true) {
-    switch (address->getKind()) {
-    default:
-      address->dump();
-      llvm_unreachable("unexpected address source.");
-
-    // Base cases: these are always the base of a formal access.
-    case ValueKind::GlobalAddrInst:
-    case ValueKind::RefElementAddrInst:
-    // An AllocBox is a fully identified memory location.
-    case ValueKind::AllocBoxInst:
-    // An AllocStack is a fully identified memory location, which may occur
-    // after inlining code already subjected to stack promotion.
-    case ValueKind::AllocStackInst:
-    // View the outer begin_access as a separate location because nested
-    // accesses do not conflict with each other.
-    case ValueKind::BeginAccessInst:
-    // A function argument is effectively a nested access, enforced
-    // independently in the caller and callee.
-    case ValueKind::SILFunctionArgument:
-    // An addressor provides access to a global or class property via a
-    // RawPointer. Calling the addressor casts that raw pointer to an address.
-    case ValueKind::PointerToAddressInst:
-      return address;
-
-    // A block argument may be a box value projected out of
-    // switch_enum. Address-type block arguments are not allowed.
-    case ValueKind::SILPHIArgument:
-      checkSwitchEnumBlockArg(cast<SILPHIArgument>(address));
-      return address;
-
-    // Load a box from an indirect payload of an opaque enum.
-    // We must have peeked past the project_box earlier in this loop.
-    // (the indirectness makes it a box, the load is for address-only).
-    // 
-    // %payload_adr = unchecked_take_enum_data_addr %enum : $*Enum, #Enum.case
-    // %box = load [take] %payload_adr : $*{ var Enum }
-    //
-    // FIXME: this case should go away with opaque values.
-    case ValueKind::LoadInst: {
-      assert(address->getType().is<SILBoxType>());
-      address = cast<LoadInst>(address)->getOperand();
-      assert(isa<UncheckedTakeEnumDataAddrInst>(address));
-      continue;
-    }
-    // Inductive cases:
-    // Look through address casts to find the source address.
-    case ValueKind::MarkUninitializedInst:
-    case ValueKind::OpenExistentialAddrInst:
-    case ValueKind::UncheckedAddrCastInst:
-    // Inductive cases that apply to any type.
-    case ValueKind::CopyValueInst:
-    case ValueKind::MarkDependenceInst:
-    // Look through a project_box to identify the underlying alloc_box as the
-    // accesed object. It must be possible to reach either the alloc_box or the
-    // containing enum in this loop, only looking through simple value
-    // propagation such as copy_value.
-    case ValueKind::ProjectBoxInst:
-    // Handle project_block_storage just like project_box.
-    case ValueKind::ProjectBlockStorageInst:
-    // Look through begin_borrow in case a local box is borrowed.
-    case ValueKind::BeginBorrowInst:
-      address = cast<SingleValueInstruction>(address)->getOperand(0);
-      continue;
-
-    // Subobject projections.
-    case ValueKind::StructElementAddrInst:
-    case ValueKind::TupleElementAddrInst:
-    case ValueKind::UncheckedTakeEnumDataAddrInst:
-    case ValueKind::RefTailAddrInst:
-    case ValueKind::TailAddrInst:
-    case ValueKind::IndexAddrInst:
-      address = cast<SingleValueInstruction>(address)->getOperand(0);
-      continue;
-
-    // Value to address conversions: the operand is the non-address source
-    // value. These allow local mutation of the value but should never be used
-    // for formal access of an lvalue.
-    case ValueKind::OpenExistentialBoxInst:
-    case ValueKind::ProjectExistentialBoxInst:
-    case ValueKind::ProjectValueBufferInst:
-      return SILValue();
-
-    // Local initialization: these cases are skipped.
-    case ValueKind::InitEnumDataAddrInst:
-    case ValueKind::InitExistentialAddrInst:
-    case ValueKind::AllocExistentialBoxInst:
-    case ValueKind::AllocValueBufferInst:
-    case ValueKind::SILUndef:
-      return SILValue();
-    }
-  }
-}
-
-bool swift::isPossibleFormalAccessBase(SILValue baseAddress) {
-  // A begin_access is considered a separate base for the purpose of conflict
-  // checking. However, for the purpose of inserting unenforced markers and
-  // performaing verification, it needs to be ignored.
-  while (auto *beginAccess = dyn_cast<BeginAccessInst>(baseAddress))
-    baseAddress = beginAccess->getOperand();
-
-  // Function arguments are accessed by the caller.
-  if (isa<SILFunctionArgument>(baseAddress))
-    return false;
-
-  if (isa<SILPHIArgument>(baseAddress)) {
-    checkSwitchEnumBlockArg(cast<SILPHIArgument>(baseAddress));
-    return false;
-  }
-
-  // Pointer-to-address exclusivity cannot be enforced. `baseAddress` may be
-  // pointing anywhere within an object.
-  if (isa<PointerToAddressInst>(baseAddress))
-    return false;
-
-  // Immutable values are only accessed for initialization.
-  if (isLetAccess(baseAddress))
-    return false;
-
-  // Special case unsafe value buffer access.
-  if (isa<BuiltinUnsafeValueBufferType>(
-          baseAddress->getType().getSwiftRValueType())) {
-    return false;
-  }
-  return true;
-}
 
 SILValue swift::isPartialApplyOfReabstractionThunk(PartialApplyInst *PAI) {
   if (PAI->getNumArguments() != 1)
@@ -580,221 +415,6 @@ FindClosureResult swift::findClosureForAppliedArg(SILValue V) {
     return FindClosureResult(result.PAI, true);
   }
   return FindClosureResult(PAI, false);
-}
-
-/// Helper for visitApplyAccesses that visits address-type call arguments,
-/// including arguments to @noescape functions that are passed as closures to
-/// the current call.
-static void visitApplyAccesses(ApplySite apply,
-                               std::function<void(Operand *)> visitor) {
-  for (Operand &oper : apply.getArgumentOperands()) {
-    // Consider any address-type operand an access. Whether it is read or modify
-    // depends on the argument convention.
-    if (oper.get()->getType().isAddress()) {
-      visitor(&oper);
-      continue;
-    }
-    auto fnType = oper.get()->getType().getAs<SILFunctionType>();
-    if (!fnType || !fnType->isNoEscape())
-      continue;
-
-    // When @noescape function closures are passed as arguments, their
-    // arguments are considered accessed at the call site.
-    FindClosureResult result = findClosureForAppliedArg(oper.get());
-    if (!result.PAI)
-      continue;
-
-    // Recursively visit @noescape function closure arguments.
-    visitApplyAccesses(result.PAI, visitor);
-  }
-}
-
-void swift::visitAccessedAddress(SILInstruction *I,
-                                 std::function<void(Operand *)> visitor) {
-  assert(I->mayReadOrWriteMemory());
-
-  // Reference counting instructions do not access user visible memory.
-  if (isa<RefCountingInst>(I))
-    return;
-
-  if (isa<DeallocationInst>(I))
-    return;
-
-  if (auto apply = FullApplySite::isa(I)) {
-    visitApplyAccesses(apply, visitor);
-    return;
-  }
-
-  if (auto builtin = dyn_cast<BuiltinInst>(I)) {
-    if (auto Kind = builtin->getBuiltinKind()) {
-      switch (Kind.getValue()) {
-      default:
-        I->dump();
-        llvm_unreachable("unexpected bulitin memory access.");
-
-      // Buitins that affect memory but can't be formal accesses.
-      case BuiltinValueKind::UnexpectedError:
-      case BuiltinValueKind::ErrorInMain:
-      case BuiltinValueKind::IsOptionalType:
-      case BuiltinValueKind::AllocRaw:
-      case BuiltinValueKind::DeallocRaw:
-      case BuiltinValueKind::Fence:
-      case BuiltinValueKind::StaticReport:
-      case BuiltinValueKind::Once:
-      case BuiltinValueKind::OnceWithContext:
-      case BuiltinValueKind::Unreachable:
-      case BuiltinValueKind::CondUnreachable:
-      case BuiltinValueKind::DestroyArray:
-      case BuiltinValueKind::UnsafeGuaranteed:
-      case BuiltinValueKind::UnsafeGuaranteedEnd:
-      case BuiltinValueKind::Swift3ImplicitObjCEntrypoint:
-      case BuiltinValueKind::TSanInoutAccess:
-        return;
-
-      // General memory access to a pointer in first operand position.
-      case BuiltinValueKind::CmpXChg:
-      case BuiltinValueKind::AtomicLoad:
-      case BuiltinValueKind::AtomicStore:
-      case BuiltinValueKind::AtomicRMW:
-        // Currently ignored because the access is on a RawPointer, not a
-        // SIL address.
-        // visitor(&builtin->getAllOperands()[0]);
-        return;
-
-      // Arrays: (T.Type, Builtin.RawPointer, Builtin.RawPointer,
-      // Builtin.Word)
-      case BuiltinValueKind::CopyArray:
-      case BuiltinValueKind::TakeArrayNoAlias:
-      case BuiltinValueKind::TakeArrayFrontToBack:
-      case BuiltinValueKind::TakeArrayBackToFront:
-      case BuiltinValueKind::AssignCopyArrayNoAlias:
-      case BuiltinValueKind::AssignCopyArrayFrontToBack:
-      case BuiltinValueKind::AssignCopyArrayBackToFront:
-      case BuiltinValueKind::AssignTakeArray:
-        // Currently ignored because the access is on a RawPointer.
-        // visitor(&builtin->getAllOperands()[1]);
-        // visitor(&builtin->getAllOperands()[2]);
-        return;
-      }
-    }
-    if (auto ID = builtin->getIntrinsicID()) {
-      switch (ID.getValue()) {
-      // Exhaustively verifying all LLVM instrinsics that access memory is
-      // impractical. Instead, we call out the few common cases and return in
-      // the default case.
-      default:
-        return;
-      case llvm::Intrinsic::memcpy:
-      case llvm::Intrinsic::memmove:
-        // Currently ignored because the access is on a RawPointer.
-        // visitor(&builtin->getAllOperands()[0]);
-        // visitor(&builtin->getAllOperands()[1]);
-        return;
-      case llvm::Intrinsic::memset:
-        // Currently ignored because the access is on a RawPointer.
-        // visitor(&builtin->getAllOperands()[0]);
-        return;
-      }
-    }
-    llvm_unreachable("Must be either a builtin or intrinsic.");
-  }
-
-  switch (I->getKind()) {
-  default:
-    I->dump();
-    llvm_unreachable("unexpected memory access.");
-
-  case SILInstructionKind::AssignInst:
-    visitor(&I->getAllOperands()[AssignInst::Dest]);
-    return;
-
-  case SILInstructionKind::CheckedCastAddrBranchInst:
-    visitor(&I->getAllOperands()[CheckedCastAddrBranchInst::Src]);
-    visitor(&I->getAllOperands()[CheckedCastAddrBranchInst::Dest]);
-    return;
-
-  case SILInstructionKind::CopyAddrInst:
-    visitor(&I->getAllOperands()[CopyAddrInst::Src]);
-    visitor(&I->getAllOperands()[CopyAddrInst::Dest]);
-    return;
-
-  case SILInstructionKind::StoreInst:
-  case SILInstructionKind::StoreBorrowInst:
-  case SILInstructionKind::StoreUnownedInst:
-  case SILInstructionKind::StoreWeakInst:
-    visitor(&I->getAllOperands()[StoreInst::Dest]);
-    return;
-
-  case SILInstructionKind::SelectEnumAddrInst:
-    visitor(&I->getAllOperands()[0]);
-    return;
-
-  case SILInstructionKind::InitExistentialAddrInst:
-  case SILInstructionKind::InjectEnumAddrInst:
-  case SILInstructionKind::LoadInst:
-  case SILInstructionKind::LoadBorrowInst:
-  case SILInstructionKind::LoadWeakInst:
-  case SILInstructionKind::LoadUnownedInst:
-  case SILInstructionKind::OpenExistentialAddrInst:
-  case SILInstructionKind::SwitchEnumAddrInst:
-  case SILInstructionKind::UncheckedTakeEnumDataAddrInst:
-  case SILInstructionKind::UnconditionalCheckedCastInst: {
-    assert(I->getNumOperands() - I->getNumTypeDependentOperands() == 1);
-    Operand *singleOperand = &I->getAllOperands()[0];
-    if (singleOperand->get()->getType().isAddress())
-      visitor(singleOperand);
-    return;
-  }
-  // Non-access cases: these are marked with memory side effects, but, by
-  // themselves, do not access formal memory.
-  case SILInstructionKind::AbortApplyInst:
-  case SILInstructionKind::AllocBoxInst:
-  case SILInstructionKind::AllocExistentialBoxInst:
-  case SILInstructionKind::AllocGlobalInst:
-  case SILInstructionKind::BeginAccessInst:
-  case SILInstructionKind::BeginApplyInst:
-  case SILInstructionKind::BeginBorrowInst:
-  case SILInstructionKind::BeginUnpairedAccessInst:
-  case SILInstructionKind::BindMemoryInst:
-  case SILInstructionKind::CheckedCastValueBranchInst:
-  case SILInstructionKind::CondFailInst:
-  case SILInstructionKind::CopyBlockInst:
-  case SILInstructionKind::CopyValueInst:
-  case SILInstructionKind::CopyUnownedValueInst:
-  case SILInstructionKind::DeinitExistentialAddrInst:
-  case SILInstructionKind::DeinitExistentialValueInst:
-  case SILInstructionKind::DestroyAddrInst:
-  case SILInstructionKind::DestroyValueInst:
-  case SILInstructionKind::EndAccessInst:
-  case SILInstructionKind::EndApplyInst:
-  case SILInstructionKind::EndBorrowArgumentInst:
-  case SILInstructionKind::EndBorrowInst:
-  case SILInstructionKind::EndUnpairedAccessInst:
-  case SILInstructionKind::EndLifetimeInst:
-  case SILInstructionKind::ExistentialMetatypeInst:
-  case SILInstructionKind::FixLifetimeInst:
-  case SILInstructionKind::InitExistentialValueInst:
-  case SILInstructionKind::IsUniqueInst:
-  case SILInstructionKind::IsEscapingClosureInst:
-  case SILInstructionKind::IsUniqueOrPinnedInst:
-  case SILInstructionKind::KeyPathInst:
-  case SILInstructionKind::OpenExistentialBoxInst:
-  case SILInstructionKind::OpenExistentialBoxValueInst:
-  case SILInstructionKind::OpenExistentialValueInst:
-  case SILInstructionKind::PartialApplyInst:
-  case SILInstructionKind::ProjectValueBufferInst:
-  case SILInstructionKind::StrongPinInst:
-  case SILInstructionKind::YieldInst:
-  case SILInstructionKind::UnwindInst:
-  case SILInstructionKind::UncheckedOwnershipConversionInst:
-  case SILInstructionKind::UncheckedRefCastAddrInst:
-  case SILInstructionKind::UnconditionalCheckedCastAddrInst:
-  case SILInstructionKind::UnconditionalCheckedCastValueInst:
-  case SILInstructionKind::UnownedReleaseInst:
-  case SILInstructionKind::UnownedRetainInst:
-  case SILInstructionKind::ValueMetatypeInst:
-    return;
-  }
 }
 
 namespace {
